Move comments to the top of the function bodies
Move comments about functions to the line before the functions so that
we don't need to look at the code to see what is going on.

Many of these comments need to be improved, detailing what kind of data
the functions receive, what they return and what they do more precisely.


git-svn-id: file:///tmp/orig/avr-evtd/trunk@153 a2d00c94-3d26-0410-a74b-ce61611494ea