Add some eclipse files for refactoring.
I don't usually use eclipse, but it is one of the tools that I know how to
use to refactor the code slightly and the current state of the project is so
crufty that anything that can help with an understanding of the code is
welcome (or else, I would have rewritten this from scratch).

In fact, I'm this close -><- of throwing up everything to the trash and,
indeed, refactoring it to be more readable.

Using a third-party library for reading the configuration file is something
that I have thought about, since it would remove the need of implementing a
parser/finite state machine, but as this program is meant to be for
emergency situations, it would be better to have everything self-contained.